requirements.txt
# Core Dependencies
mysql-connector-python==8.0.33
python-dotenv==1.0.0

# GUI
customtkinter==5.2.0
ttkthemes==3.2.2
Pillow==10.0.0

# PDF Generation
reportlab==4.0.4

# Excel Handling
openpyxl==3.1.2

# Presentations
python-pptx==0.6.23 # Agregado para manejar archivos .pptx

# Email
email-validator==2.0.0

# Development
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6

# Documentation
sphinx==7.0.1